% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rgl.R
\name{animate_scene}
\alias{animate_scene}
\title{A few demo animation routines}
\usage{
animate_scene(routine = 1, duration = 6)
}
\arguments{
\item{routine}{default: 1}

\item{duration}{duration. default 5}
}
\description{
This is just a wrapper around \code{rgl} functions \code{play3d()},
\code{spin3d()} and \code{par3dinter()}
}
