% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2hex}
\alias{col2hex}
\title{Convert a devices colour back into an RGB colour}
\usage{
col2hex(col)
}
\arguments{
\item{col}{4 element RGBA colour with values in range [0-255]}
}
\description{
Convert a devices colour back into an RGB colour
}
