% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl-raster-png.R
\name{convert_device_raster_to_array}
\alias{convert_device_raster_to_array}
\title{Convert a device raster to an array}
\usage{
convert_device_raster_to_array(device_raster, w, h)
}
\arguments{
\item{device_raster}{raster passed into device. array of ints which
contain ABGR pixels packed 4-bytes-to-an-int}

\item{w, h}{data dimensions}
}
\description{
Convert a device raster to an array
}
