% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl-raster-png.R
\name{convert_device_raster_to_png}
\alias{convert_device_raster_to_png}
\title{Helper function to convert device raster data to a PNG file}
\usage{
convert_device_raster_to_png(args, flip_y = FALSE, debug_items = c())
}
\arguments{
\item{args}{arguments to a 'raster' device call}

\item{flip_y}{whether to flip the image vertically. default: FALSE}

\item{debug_items}{character vector of names of things to debug}
}
\value{
list with 'png_filename' and the RGBA 'array_data' (values in [0, 1])
}
\description{
Helper function to convert device raster data to a PNG file
}
