% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-font-as-image.R
\name{create_text_as_png}
\alias{create_text_as_png}
\title{Render a text string to a PNG file}
\usage{
create_text_as_png(
  text,
  font_size = 30,
  font = "sans",
  colour = "black",
  angle = 0
)
}
\arguments{
\item{text}{text}

\item{font_size}{font size}

\item{font}{'sans'}

\item{colour}{'black'}

\item{angle}{0}
}
\value{
list of meta-information about the png. width, height, png_filename
}
\description{
Render a text string to a PNG file
}
