% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl-callback.R
\name{rgl_callback}
\alias{rgl_callback}
\title{ASCII callback for the rdevice}
\usage{
rgl_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device function call}

\item{args}{arguments to device function call}

\item{state}{list of rdata, dd and gc. Some or all of which may be NULL}
}
\description{
ASCII callback for the rdevice
}
