% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-rgl-doc.R
\name{RGLDoc}
\alias{RGLDoc}
\title{Class for creating an opengl scene with rgl.}
\description{
Class for creating an opengl scene with rgl.

Class for creating an opengl scene with rgl.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{default_material}}{default material}

\item{\code{width, height}}{dimensions of rgl window}

\item{\code{polygon_offset, polygon_offset_delta}}{current polygon offset, and
the delta to adjust the offset by every time an element is drawn}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{width, height}}{dimensions of rgl window}

\item{\code{polygon_offset, polygon_offset_delta}}{current polygon offset, and
the delta to adjust the offset by every time an element is drawn}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RGLDoc$new()}}
\item \href{#method-circle}{\code{RGLDoc$circle()}}
\item \href{#method-add_call}{\code{RGLDoc$add_call()}}
\item \href{#method-sphere}{\code{RGLDoc$sphere()}}
\item \href{#method-point}{\code{RGLDoc$point()}}
\item \href{#method-rect}{\code{RGLDoc$rect()}}
\item \href{#method-line}{\code{RGLDoc$line()}}
\item \href{#method-polyline}{\code{RGLDoc$polyline()}}
\item \href{#method-polygon_old}{\code{RGLDoc$polygon_old()}}
\item \href{#method-polygon}{\code{RGLDoc$polygon()}}
\item \href{#method-path}{\code{RGLDoc$path()}}
\item \href{#method-text}{\code{RGLDoc$text()}}
\item \href{#method-image_as_quad}{\code{RGLDoc$image_as_quad()}}
\item \href{#method-save}{\code{RGLDoc$save()}}
\item \href{#method-clone}{\code{RGLDoc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$new(
  fov = 30,
  width = 800,
  height = 600,
  zscale = 1,
  view = c("front", "flat"),
  view_angle = 0,
  view_flat_angle = -45,
  zoom = 0.8,
  view3d_args = NULL,
  light = list(theta = 0, phi = 0),
  background = list(color = "white"),
  default_material = standard_material,
  polygon_offset_delta = 0.03,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fov}}{field of view. default: 0 isometric view. Other good values are 30 and 45.}

\item{\code{width, height}}{dimensions of viewing window. default 800x600}

\item{\code{zscale}}{scale factor applied to z axis. default 1.}

\item{\code{view}}{One of 'front', 'flat'. This will be ignored if \code{view3d_args}
is set.}

\item{\code{view_angle}}{rotation angle of view in degrees. Only used if \code{view} is
'front' or 'flat'.  Default: 0}

\item{\code{view_flat_angle}}{angle to tip view when view is 'flat'.}

\item{\code{zoom}}{zoom level. default 0.800. Larger values make the plot appear
further from the screen.}

\item{\code{view3d_args}}{full viewpoint specification as a list of arguments to
\code{rgl::view3d()}.  Default: NULL}

\item{\code{light}}{list of arguments to \code{rgl.light()} to define the
light for this scene.}

\item{\code{background}}{list of arguments to \code{rgl.bg()} to define the
background for these scene}

\item{\code{default_material}}{default material for objects}

\item{\code{polygon_offset_delta}}{incremental change in \code{polygon_offset} after
every element is rendered.  Adjust this value to try and
minimise issues with z-fighting of overlapping elements.}

\item{\code{...}}{other arguments passed to \code{rgl::open3d()}
Add a circle}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-circle"></a>}}
\if{latex}{\out{\hypertarget{method-circle}{}}}
\subsection{Method \code{circle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$circle(
  x,
  y,
  z,
  r,
  colour,
  color = colour,
  alpha = 1,
  circle_segments = 30
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y, z}}{coords}

\item{\code{r}}{radius}

\item{\code{colour, color}}{colour}

\item{\code{alpha}}{alpha}

\item{\code{circle_segments}}{number of segments to render for a circle
Add an arbitrary call to this document}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_call"></a>}}
\if{latex}{\out{\hypertarget{method-add_call}{}}}
\subsection{Method \code{add_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$add_call(cc, default_args = standard_material)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cc}}{standard R call object}

\item{\code{default_args}}{default args
Add a sphere}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sphere"></a>}}
\if{latex}{\out{\hypertarget{method-sphere}{}}}
\subsection{Method \code{sphere()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$sphere(x, y, z, r, colour, color = colour, alpha = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y, z}}{coords}

\item{\code{r}}{radius}

\item{\code{colour, color}}{colour}

\item{\code{alpha}}{alpha
Add a point}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-point"></a>}}
\if{latex}{\out{\hypertarget{method-point}{}}}
\subsection{Method \code{point()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$point(x, y, z, r, colour, color = colour, alpha = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y, z}}{coords}

\item{\code{r}}{radius}

\item{\code{colour, color}}{colour}

\item{\code{alpha}}{alpha
Add a rect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rect"></a>}}
\if{latex}{\out{\hypertarget{method-rect}{}}}
\subsection{Method \code{rect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$rect(
  x1,
  y1,
  x2,
  y2,
  z,
  colour,
  fill,
  lwd = 3,
  color = colour,
  alpha = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2, z}}{coords}

\item{\code{colour, color, fill}}{colour}

\item{\code{lwd}}{line width i.e. size}

\item{\code{alpha}}{alpha
Add a line}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-line"></a>}}
\if{latex}{\out{\hypertarget{method-line}{}}}
\subsection{Method \code{line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$line(x1, y1, x2, y2, z, colour, lwd = 3, color = colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2, z}}{coords}

\item{\code{colour, color}}{colour}

\item{\code{lwd}}{line width i.e. size
Add a polyline}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polyline"></a>}}
\if{latex}{\out{\hypertarget{method-polyline}{}}}
\subsection{Method \code{polyline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$polyline(xs, ys, z, colour, lwd = 3, color = colour)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys, z}}{coords}

\item{\code{colour, color}}{colour}

\item{\code{lwd}}{line width i.e. size
Add a polygon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polygon_old"></a>}}
\if{latex}{\out{\hypertarget{method-polygon_old}{}}}
\subsection{Method \code{polygon_old()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$polygon_old(xs, ys, z = 0, colour, fill, alpha = 1, lwd = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys, z}}{coords}

\item{\code{colour, fill}}{colour}

\item{\code{alpha}}{alpha}

\item{\code{lwd}}{line width i.e. size
Add a polygon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-polygon"></a>}}
\if{latex}{\out{\hypertarget{method-polygon}{}}}
\subsection{Method \code{polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$polygon(xs, ys, z = 0, colour, fill, alpha = 1, lwd = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys, z}}{coords}

\item{\code{colour, fill}}{colour}

\item{\code{alpha}}{alpha}

\item{\code{lwd}}{line width i.e. size
Add a path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path"></a>}}
\if{latex}{\out{\hypertarget{method-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$path(df, colour, fill, alpha = 1, lwd = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{polygon specification}

\item{\code{colour, fill}}{colour}

\item{\code{alpha}}{alpha}

\item{\code{lwd}}{line width i.e. size
Add text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\if{latex}{\out{\hypertarget{method-text}{}}}
\subsection{Method \code{text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$text(text, x, y, z = 0, colour, cex = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text to write}

\item{\code{x, y, z}}{coords}

\item{\code{colour, color}}{colour}

\item{\code{cex}}{character expansion values
Render an image onto a quad}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image_as_quad"></a>}}
\if{latex}{\out{\hypertarget{method-image_as_quad}{}}}
\subsection{Method \code{image_as_quad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$image_as_quad(
  png_filename,
  x,
  y,
  z,
  width,
  height,
  alpha = 1,
  angle = 0,
  option = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{png_filename}}{filename of PNG to add to image}

\item{\code{x, y}}{display position}

\item{\code{z}}{height}

\item{\code{width, height}}{display width and height}

\item{\code{alpha}}{overall alpha}

\item{\code{angle}}{rotation angle of quad around the z axis}

\item{\code{option}}{generic option to control rendering
Save as PNG or WebGL based upon suffix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$save(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{filename}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGLDoc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
