% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rgl.R
\name{save_animated_scene}
\alias{save_animated_scene}
\title{A few demo animation routines}
\usage{
save_animated_scene(
  gif_name = "ggrgl.gif",
  routine = 1,
  duration = 6,
  fps = 10
)
}
\arguments{
\item{gif_name}{name of gif output file}

\item{routine}{default: 1}

\item{duration}{duration. default 5}

\item{fps}{frames per second. default 10}
}
\description{
This is just a wrapper around \code{rgl} functions \code{play3d()},
\code{spin3d()} and \code{par3dinter()}
}
