% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-rgl-doc.R
\docType{data}
\name{standard_material}
\alias{standard_material}
\title{Standard default material}
\format{
An object of class \code{list} of length 24.
}
\usage{
standard_material
}
\description{
Standard default material
}
\keyword{datasets}
