% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot2.R
\name{theme_rgl}
\alias{theme_rgl}
\title{Theme for rgl plots which turns off some white rectangles for nicer render}
\usage{
theme_rgl(...)
}
\arguments{
\item{...}{arguments ignored}
}
\value{
some standard theme adjustments
}
\description{
Theme for rgl plots which turns off some white rectangles for nicer render
}
