% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Simpler version of 'modifyList'}
\usage{
update_list(old, new)
}
\arguments{
\item{old, new}{lists}
}
\value{
updated version of old list
}
\description{
Simpler version of 'modifyList'
}
